refactor: refactor userActivityDashboard file into smaller subcomponents
